Potential solution for issue #16
For recordings, the player now searches for any transcoded files with
the same name but different file extention (i.e. foo.mpg and foo.m4v).

All recordings are now shown in the player, however, any file that is
not .m4v, .mp4, or .mov will display "NOT TRANSCODED".

The behavior for videos (MythVideo) has not changed. The player still
screens for .m4v, .mp4, or .mov files. All other files will not be
displayed.